Refactor charts types
